# openapi/v1.yml
...
openapi: "3.0.0"
info:
  version: "1.0.0-alpha1"
  description: "API definition for Kleister, manage mod packs for Minecraft"
  title: "Kleister OpenAPI"
  contact:
    name: "Kleister"
    email: "kleister@webhippie.de"
    url: "https://kleister.eu"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://try.kleister.eu/api/v1"
security:
  - Header: []
  - Bearer: []
  - Basic: []
  - Cookie: []
paths:
...
  /fabric:
    get:
      summary: "Fetch the available Fabric versions"
      operationId: "ListFabrics"
      tags:
        - "fabric"
      parameters:
        - name: "search"
          in: "query"
          required: false
          schema:
            type: "string"
          description: "Search query"
          x-example: "John Doe"
      responses:
        "200":
          description: "A collection of fabric versions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fabrics"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    put:
      summary: "Update the available Fabric versions"
      operationId: "UpdateFabric"
      tags:
        - "fabric"
      responses:
        "200":
          description: "Plain success message"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "503":
          description: "If remote source is not available"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
  /fabric/{fabric_id}/builds:
    get:
      summary: "Fetch the builds attached to a Fabric version"
      operationId: "ListFabricBuilds"
      tags:
        - "fabric"
      parameters:
        - in: "path"
          name: "fabric_id"
          description: "A fabric identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "2oexcyrows1j0bo0itju"
        - name: "search"
          in: "query"
          required: false
          schema:
            type: "string"
          description: "Search query"
          x-example: "John Doe"
        - name: "sort"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "build_name"
            enum:
              - "build_name"
              - "build_public"
              - "pack_slug"
              - "pack_name"
          description: "Sorting column"
          x-example: "name"
        - name: "order"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "asc"
            enum:
              - "asc"
              - "desc"
          description: "Sorting order"
          x-example: "desc"
        - name: "limit"
          in: "query"
          required: false
          schema:
            type: "integer"
            default: 100
          description: "Paging limit"
          x-example: 100
        - name: "offset"
          in: "query"
          required: false
          schema:
            type: "integer"
            default: 0
          description: "Paging offset"
          x-example: 0
      responses:
        "200":
          description: "A collection of attached builds"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fabric_builds"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Fabric or build not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    post:
      summary: "Attach a build to a Fabric version"
      operationId: "AttachFabricToBuild"
      tags:
        - "fabric"
      parameters:
        - in: "path"
          name: "fabric_id"
          description: "A fabric identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "2oexcyrows1j0bo0itju"
      requestBody:
        description: "The build data to attach"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/fabric_build_params"
      responses:
        "200":
          description: "Plain success message"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Fabric or build not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "412":
          description: "Build is already attached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "422":
          description: "Failed to validate request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    delete:
      summary: "Unlink a build from a Fabric version"
      operationId: "DeleteFabricFromBuild"
      tags:
        - "fabric"
      parameters:
        - in: "path"
          name: "fabric_id"
          description: "A fabric identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "2oexcyrows1j0bo0itju"
      requestBody:
        description: "The build data to unlink"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/fabric_build_params"
      responses:
        "200":
          description: "Plain success message"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Fabric or build not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "412":
          description: "Build is not attached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
...
  /packs/{pack_id}/builds:
    get:
      summary: "Fetch all available builds for a pack"
      operationId: "ListBuilds"
      tags:
        - "pack"
      parameters:
        - in: "path"
          name: "pack_id"
          description: "A pack identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "edtmbzwztudkdhhfemhd"
        - name: "search"
          in: "query"
          required: false
          schema:
            type: "string"
          description: "Search query"
          x-example: "John Doe"
        - name: "sort"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "name"
            enum:
              - "name"
              - "public"
          description: "Sorting column"
          x-example: "name"
        - name: "order"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "asc"
            enum:
              - "asc"
              - "desc"
          description: "Sorting order"
          x-example: "desc"
        - name: "limit"
          in: "query"
          required: false
          schema:
            type: "integer"
          description: "Paging limit"
          x-example: 100
        - name: "offset"
          in: "query"
          required: false
          schema:
            type: "integer"
          description: "Paging offset"
          x-example: 0
      responses:
        "200":
          description: "A collection of builds"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/builds"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Pack not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    post:
      summary: "Create a new build for a pack"
      operationId: "CreateBuild"
      tags:
        - "pack"
      parameters:
        - in: "path"
          name: "pack_id"
          description: "A pack identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "edtmbzwztudkdhhfemhd"
      requestBody:
        description: "The build data to create"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/build"
      responses:
        "200":
          description: "The created build build"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/build"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Pack not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "422":
          description: "Failed to validate request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
...
components:
  securitySchemes:
    Header:
      type: apiKey
      in: header
      name: X-API-Key
    Bearer:
      type: http
      scheme: bearer
    Basic:
      type: http
      scheme: basic
    Cookie:
      type: apiKey
      in: header
      name: Cookie
  schemas:
...
    minecraft:
      title: "Minecraft"
      description: "Model to represent minecraft"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        type:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    forge:
      title: "Forge"
      description: "Model to represent forge"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        minecraft:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    neoforge:
      title: "Neoforge"
      description: "Model to represent neoforge"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    quilt:
      title: "Quilt"
      description: "Model to represent quilt"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    fabrics:
      title: "Fabrics"
      description: "Model to represent list of fabrics"
      type: "object"
      properties:
        total:
          type: integer
          format: int64
        versions:
          type: "array"
          items:
            $ref: "#/components/schemas/fabric"
    fabric:
      title: "Fabric"
      description: "Model to represent fabric"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
    fabric_builds:
      title: "Fabric Builds"
      description: "Model to represent fabric builds"
      type: "object"
      properties:
        fabric:
          $ref: "#/components/schemas/fabric"
          readOnly: true
        total:
          type: integer
          format: int64
        builds:
          type: "array"
          items:
            $ref: "#/components/schemas/build"
    fabric_build_params:
      title: "Fabric Build Params"
      description: "Model to represent params for fabric builds"
      type: "object"
      required:
        - "pack"
        - "build"
      properties:
        pack:
          type: "string"
        build:
          type: "string"
...
    pack:
      title: "Pack"
      description: "Model to represent pack"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        icon:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/pack_icon"
        logo:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/pack_logo"
        back:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/pack_back"
        slug:
          type: "string"
          x-omitempty: true
          x-nullable: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        website:
          type: "string"
          x-omitempty: true
          x-nullable: true
        public:
          type: "boolean"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
    pack_icon:
      title: "Pack Icon"
      description: "Model to represent pack icon"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        slug:
          type: "string"
          x-omitempty: true
          x-nullable: true
        content_type:
          type: "string"
          x-omitempty: true
          x-nullable: true
        md5:
          type: "string"
          x-omitempty: true
          x-nullable: true
        path:
          type: "string"
          x-omitempty: true
          x-nullable: true
        url:
          type: "string"
          x-omitempty: true
          x-nullable: true
        upload:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
    pack_logo:
      title: "Pack Logo"
      description: "Model to represent pack logo"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        slug:
          type: "string"
          x-omitempty: true
          x-nullable: true
        content_type:
          type: "string"
          x-omitempty: true
          x-nullable: true
        md5:
          type: "string"
          x-omitempty: true
          x-nullable: true
        path:
          type: "string"
          x-omitempty: true
          x-nullable: true
        url:
          type: "string"
          x-omitempty: true
          x-nullable: true
        upload:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
    pack_back:
      title: "Pack Background"
      description: "Model to represent pack background"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        slug:
          type: "string"
          x-omitempty: true
          x-nullable: true
        content_type:
          type: "string"
          x-omitempty: true
          x-nullable: true
        md5:
          type: "string"
          x-omitempty: true
          x-nullable: true
        path:
          type: "string"
          x-omitempty: true
          x-nullable: true
        url:
          type: "string"
          x-omitempty: true
          x-nullable: true
        upload:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
    builds:
      title: "Builds"
      description: "Model to represent list of builds"
      type: "object"
      properties:
        total:
          type: integer
          format: int64
        pack:
          readOnly: true
          $ref: "#/components/schemas/pack"
        builds:
          type: "array"
          items:
            $ref: "#/components/schemas/build"
    build:
      title: "Build"
      description: "Model to represent build"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        pack:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/pack"
        minecraft_id:
          type: "string"
          x-omitempty: true
          x-nullable: true
        minecraft:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/minecraft"
        forge_id:
          type: "string"
          x-omitempty: true
          x-nullable: true
        forge:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/forge"
        neoforge_id:
          type: "string"
          x-omitempty: true
          x-nullable: true
        neoforge:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/neoforge"
        quilt_id:
          type: "string"
          x-omitempty: true
          x-nullable: true
        quilt:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/quilt"
        fabric_id:
          type: "string"
          x-omitempty: true
          x-nullable: true
        fabric:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/fabric"
        slug:
          type: "string"
          x-omitempty: true
          x-nullable: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        java:
          type: "string"
          x-omitempty: true
          x-nullable: true
        memory:
          type: "string"
          x-omitempty: true
          x-nullable: true
        latest:
          type: "boolean"
          x-omitempty: true
          x-nullable: true
        recommended:
          type: "boolean"
          x-omitempty: true
          x-nullable: true
        public:
          type: "boolean"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    notification:
      title: "Notification"
      description: "Generic response for errors and validations"
      type: "object"
      properties:
        status:
          type: "integer"
          x-omitempty: true
        message:
          type: "string"
          x-omitempty: true
        errors:
          type: "array"
          items:
            $ref: "#/components/schemas/validation"
    validation:
      title: "Validation"
      description: "General structure to show validation errors"
      type: "object"
      properties:
        field:
          type: "string"
        message:
          type: "string"
...